requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
asyncpg==0.29.0
outline-vpn-api==6.3.0
python-decouple==3.8
YooMoney==0.1.1
paramiko==3.5.0
asyncssh==2.19.0
yookassa==3.4.3
pyCryptomusAPI==0.1.0